client_py/requirements.txt
altgraph==0.17.4
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
packaging==25.0
pefile==2023.2.7
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyperclip==1.9.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pywin32-ctypes==0.2.3
requests==2.32.4
urllib3==2.5.0
grpcio
grpcio-tools
tenacity